Add new parameters to IPrecision annotation. One parameter to specify a class which contains a method to execute when the battery level changes and a double array to specify the parameters used by the method to execute

+++ app/src/main/java/com/packetsniffer/emenegal/packetsniffer/api/strategy/Util.java
@@ -61,7 +61,7 @@ public class Util {
         }else{
             for(Field field : fields) {
                 try {
-                    field.setInt(field.getClass(), field.getAnnotation(IPrecision.class).lower());
+                    field.setDouble(field.getClass(), field.getAnnotation(IPrecision.class).lower());
                 } catch (IllegalAccessException e) {
                     e.printStackTrace();
                 }
